move ResolvableChainConfig to `chain_config.rs`
